test(shell): split the daemon bash-wrapper suite out of shell-ready.test.ts
CI's static-analysis gate failed: shell-ready.test.ts counted 801 non-blank,
non-comment lines against a 800-line cap for tests. The file was already at the
cap on main; resolving the wrapper root instead of hardcoding it made a few
call sites longer, and the formatter wrapped them.

The repo forbids disabling or bumping max-lines, so split rather than silence.
The bash-wrapper tests -- OSC 133 lifecycle, PROMPT_COMMAND composition, the
live-bash rcfile runs -- are a self-contained group, and the live-shell harness
they share with the launch-config suite moves to a fixture module rather than
being duplicated, so the two copies cannot drift.

Same 29 tests before and after.
