+++ src/main/daemon/shell-ready.test.ts
@@ -3,7 +3,6 @@ import { spawnSync } from 'node:child_process'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
 import { existsSync, mkdirSync, mkdtempSync, readFileSync, rmSync, writeFileSync } from 'node:fs'
-import type * as DaemonBashRcfileModule from './daemon-bash-shell-ready-rcfile'
 import {
   OVERLAY_ONLY_FEATURES,
   STARTUP_COMMAND_FEATURES
@@ -32,14 +31,7 @@ async function importFreshShellReady() {
   }
 }
 
-async function importFreshDaemonBashRcfile(): Promise<typeof DaemonBashRcfileModule> {
-  vi.resetModules()
-  return import('./daemon-bash-shell-ready-rcfile')
-}
-
 const describePosix = process.platform === 'win32' ? describe.skip : describe
-const hasBash = process.platform !== 'win32' && spawnSync('bash', ['--version']).status === 0
-const itWithBash = hasBash ? it : it.skip
 const hasZsh = process.platform !== 'win32' && spawnSync('zsh', ['--version']).status === 0
 const itWithZsh = hasZsh ? it : it.skip
 const FISH = resolveFishBinary()
@@ -138,46 +130,6 @@ async function runInteractiveZshRc(args: {
   clearTimeout(deadline)
   proc.kill()
   return output
-}
-
-function runInteractiveBashRcfile(rcfileContent: string, tempDir: string): string {
-  const rcfile = join(tempDir, 'bash-osc133-rcfile')
-  writeFileSync(rcfile, rcfileContent)
-
-  const result = spawnSync(
-    'bash',
-    ['-lc', 'bash --noprofile --rcfile "$1" -i 2>&1', 'bash', rcfile],
-    {
-      input: 'true\nfalse\nexit 0\n',
-      encoding: 'utf8',
-      env: {
-        ...process.env,
-        HOME: tempDir,
-        ORCA_SHELL_FEATURES: 'ready',
-        TERM: process.env.TERM || 'xterm'
-      },
-      timeout: 5000
-    }
-  )
-
-  expect(result.error).toBeUndefined()
-  expect(result.status).toBe(0)
-  return result.stdout
-}
-
-function expectBashOsc133Lifecycle(output: string): void {
-  const oscA = '\x1b]133;A\x07'
-  const oscC = '\x1b]133;C\x07'
-  const oscD = '\x1b]133;D;'
-  const firstPromptMarker = output.indexOf(oscA)
-
-  expect(firstPromptMarker).toBeGreaterThanOrEqual(0)
-  expect(output.slice(0, firstPromptMarker)).not.toContain(oscC)
-  expect(output.slice(0, firstPromptMarker)).not.toContain(oscD)
-  expect(output).toContain(`${oscD}0\x07${oscA}`)
-  expect(output).toContain(`${oscD}1\x07${oscA}`)
-  expect(output.split(oscC)).toHaveLength(4)
-  expect(output.split(oscD)).toHaveLength(3)
 }
 
 function expectZdotdirSourceContext(content: string, fileName: '.zprofile' | '.zshrc' | '.zlogin') {
@@ -684,143 +636,6 @@ describePosix('daemon shell-ready launch config', () => {
       expect(wrapperFile).not.toContain('ORCA_PRIME_AGENT_STATUS_EXTENSION')
       expect(wrapperFile).not.toContain('command prime-agent --extension')
     }
-  })
-
-  // Why: regression guard for issue #2422 — bash wrapper must emit OSC 133 C/D so SSH sessions clear stale 'working' agent rows.
-  it('emits OSC 133 C/D markers in the daemon bash wrapper', async () => {
-    const { getShellReadyLaunchConfig } = await importFreshShellReady()
-
-    getShellReadyLaunchConfig('/bin/zsh')
-    getShellReadyLaunchConfig('/bin/bash')
-
-    // Why .zshenv: the zsh markers live in the epilogue, behind `markers`.
-    const zshrc = readFileSync(join(getShellReadyWrapperRoot(), 'zsh', '.zshenv'), 'utf8')
-    const bashRc = readFileSync(join(getShellReadyWrapperRoot(), 'bash', 'rcfile'), 'utf8')
-
-    expect(bashRc).toContain('printf "\\033]133;D;%s\\007"')
-    expect(bashRc).toContain('printf "\\033]777;orca-shell-start:%s\\007" "$$"')
-    expect(bashRc).toContain('printf "\\033]133;C\\007"')
-    expect(bashRc).toContain('__orca_prepend_prompt_command "__orca_osc133_precmd"')
-    expect(bashRc).toContain('__orca_append_prompt_command "__orca_osc133_epilogue"')
-    // DEBUG is armed after PROMPT_COMMAND setup so rcfile commands aren't seen as foreground; lastIndexOf skips the epilogue's re-arm.
-    expect(bashRc.lastIndexOf("trap '__orca_osc133_preexec' DEBUG")).toBeGreaterThan(
-      bashRc.indexOf('__orca_append_prompt_command "__orca_osc133_epilogue"')
-    )
-    expect(zshrc).toContain('printf "\\033]133;D;%s\\007"')
-    expect(zshrc).toContain('printf "\\033]133;C\\007"')
-  })
-
-  itWithBash(
-    'runs the daemon bash wrapper without fake C/D markers before the first prompt',
-    async () => {
-      const { getDaemonBashShellReadyRcfileContent } = await importFreshDaemonBashRcfile()
-
-      const output = runInteractiveBashRcfile(getDaemonBashShellReadyRcfileContent(), userDataPath)
-
-      expectBashOsc133Lifecycle(output)
-    }
-  )
-
-  itWithBash(
-    'preserves prompt hooks and existing DEBUG traps without fake command markers',
-    async () => {
-      const { getDaemonBashShellReadyRcfileContent } = await importFreshDaemonBashRcfile()
-      writeFileSync(
-        join(userDataPath, '.bash_profile'),
-        [
-          'PROMPT_COMMAND=\'AFTER_FIRST_PROMPT=1; printf "PROMPT_HOOK\\n"\'',
-          'trap \'if [[ -n "${AFTER_FIRST_PROMPT:-}" ]]; then\n  printf "USER_DEBUG_AFTER\\n"\nfi\' DEBUG'
-        ].join('\n')
-      )
-
-      const output = runInteractiveBashRcfile(getDaemonBashShellReadyRcfileContent(), userDataPath)
-
-      expect(output).toContain('PROMPT_HOOK')
-      expect(output).toContain('USER_DEBUG_AFTER')
-      expectBashOsc133Lifecycle(output)
-    }
-  )
-
-  itWithBash(
-    'still emits 133;C when bash-preexec re-arms the DEBUG trap at first prompt',
-    async () => {
-      const { getDaemonBashShellReadyRcfileContent } = await importFreshDaemonBashRcfile()
-      // Minimal bash-preexec imitation: re-arms its own DEBUG trap from PROMPT_COMMAND at first prompt, silencing Orca's trap.
-      writeFileSync(
-        join(userDataPath, '.bash_profile'),
-        [
-          'preexec_functions=()',
-          '__bp_preexec_invoke_exec() {',
-          '  [[ -n "${__bp_interactive_mode:-}" ]] || return',
-          '  __bp_interactive_mode=""',
-          '  local f',
-          '  for f in "${preexec_functions[@]}"; do "$f" "$BASH_COMMAND"; done',
-          '}',
-          "__bp_arm() { __bp_interactive_mode=1; trap '__bp_preexec_invoke_exec' DEBUG; }",
-          'PROMPT_COMMAND="${PROMPT_COMMAND:+$PROMPT_COMMAND;}__bp_arm"'
-        ].join('\n')
-      )
-
-      const output = runInteractiveBashRcfile(getDaemonBashShellReadyRcfileContent(), userDataPath)
-
-      expectBashOsc133Lifecycle(output)
-    }
-  )
-
-  itWithBash(
-    'dispatches a non-empty preexec_functions against the real command, not Orca hooks',
-    async () => {
-      const { getDaemonBashShellReadyRcfileContent } = await importFreshDaemonBashRcfile()
-      // Why: the epilogue chains bash-preexec's re-armed DEBUG trap, so a real preexec callback must fire against the user's command.
-      writeFileSync(
-        join(userDataPath, '.bash_profile'),
-        [
-          'preexec_functions=(__user_preexec)',
-          '__user_preexec() { printf \'USER_PREEXEC:%s\\n\' "$1"; }',
-          '__bp_inside=0',
-          '__bp_last_hist=""',
-          '__bp_preexec_invoke_exec() {',
-          '  (( __bp_inside > 0 )) && return',
-          '  [[ -n "${__bp_interactive_mode:-}" ]] || return',
-          '  local __bp_inside=1',
-          '  local this_command',
-          '  this_command="$(builtin history 1)"',
-          '  this_command="${this_command#"${this_command%%[![:space:]]*}"}"',
-          '  this_command="${this_command#* }"',
-          '  this_command="${this_command#"${this_command%%[![:space:]]*}"}"',
-          '  [[ -n "$this_command" && "$this_command" != "$__bp_last_hist" ]] || return',
-          '  __bp_last_hist="$this_command"',
-          '  __bp_interactive_mode=""',
-          '  local f',
-          '  for f in "${preexec_functions[@]}"; do "$f" "$this_command"; done',
-          '}',
-          "__bp_arm() { set -o functrace; __bp_interactive_mode=1; trap '__bp_preexec_invoke_exec' DEBUG; }",
-          'PROMPT_COMMAND="${PROMPT_COMMAND:+$PROMPT_COMMAND;}__bp_arm"'
-        ].join('\n')
-      )
-
-      const output = runInteractiveBashRcfile(getDaemonBashShellReadyRcfileContent(), userDataPath)
-
-      expectBashOsc133Lifecycle(output)
-      expect(output).toContain('USER_PREEXEC:true')
-      expect(output).toContain('USER_PREEXEC:false')
-      expect(output).not.toContain('USER_PREEXEC:__orca_osc133')
-      expect(output).not.toContain('USER_PREEXEC:__bp_')
-    }
-  )
-
-  itWithBash('normalizes array PROMPT_COMMAND hooks so bash 3.2 still runs cleanup', async () => {
-    const { getDaemonBashShellReadyRcfileContent } = await importFreshDaemonBashRcfile()
-    writeFileSync(
-      join(userDataPath, '.bash_profile'),
-      'PROMPT_COMMAND=(\'printf "PROMPT_ARRAY_A\\n"\' \'printf "PROMPT_ARRAY_B\\n";  \')\n'
-    )
-
-    const output = runInteractiveBashRcfile(getDaemonBashShellReadyRcfileContent(), userDataPath)
-
-    expect(output.split('PROMPT_ARRAY_A')).toHaveLength(4)
-    expect(output.split('PROMPT_ARRAY_B')).toHaveLength(4)
-    expectBashOsc133Lifecycle(output)
   })
 
   it('preserves a real inherited ZDOTDIR as ORCA_ORIG_ZDOTDIR', async () => {
